+++ src/types/v2/tweet.definition.v2.ts
@@ -1,15 +1,5 @@
 import type { UserV2 } from './user.v2.types';
 
-export interface ErrorV2 {
-  value?: string;
-  detail: string;
-  title: string;
-  resource_type?: string;
-  parameter?: string;
-  resource_id?: string;
-  type: string;
-}
-
 export interface PlaceV2 {
   full_name: string;
   id: string;
